alumni_website/requirements.txt
django-htmx
django-daphne (python -m pip install -U 'channels[daphne]')
shortuuid